backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
pillow==12.1.0
python-decouple==3.8
sqlparse==0.5.5
psycopg2-binary==2.9.10
dj-database-url==2.3.0
